Fix Bindings renderring for admin
With new neutronclient objects are immutable to changes, thus writing
down properties like binding__* does not work like expected.

In the meanwhile regular binding_* are available for ports and
populated properly by the client.

This patch switching to usage of such properties for the template only
to make the patch backportable. Further rehauwling and renaming
of properties is expected to be done in follow-up patches.

Closes-Bug: #2095150
Change-Id: I57cba985d427507a4177e9b919cfae0a8188395a